trading_concierge_mo/mcp/requirements.txt
mcp>=1.26.0
httpx>=0.28.1
fastapi>=0.128.0
uvicorn>=0.40.0
pydantic>=2.12.5
pytest==8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.2.1
urllib3>=2.2.5
